rename and split dts
The convention for sharing common features for different platforms is to
use .dtsi for describing commonalities, then .dts files for concrete
platforms.  Split the .dts file into the .dtsi, but move the cpus,
fcram, and bootargs into the machine specific .dts file.

This should enable a simple nintendo3ds_ktr.dts file to be added.

To build:
$ ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- make \
  arch/arm/boot/dts/nintendo3ds_ctr.dtb

To observe what the combined DT looks like:
$ ./scripts/dtc/dtc -O dts arch/arm/boot/dts/nintendo3ds_ctr.dtb

Fixes #12
Signed-off-by: Nick Desaulniers <nick.desaulniers@gmail.com>